Enable $compile debugging info in DEV mode (+ add a concept of dev-mode)
`$compileProvider.debugInfoEnabled(false)` makes the code faster but at
the cost of not being able to inspect scopes and bindings.

Since we don't have any concept of dev/production environments in SSUI
yet (and we can't rely on handling it in the node server since that's not used in
production), adding a simple logic that looks at the current hostname,
checking for `localhost`, `127.0.0.1` and `::1`.