Gulp task 'available-languages' to create list of available languages
By invoking

    gulp available-languages

a json file

    client/gettext/json/available_languages.json

will be created containing list of available UI languages with
corresponding locale names. This list will then be presented to
the user in the self-service UI.

The task was also added as a dependency of gulp build (for production
deployments) as well as gulp serve-dev (for development setup).

Optionally, it is possible to specify a json file

    client/gettext/json/supported_languages.json

containing an array of supported languages which will restrict the list
of UI languages to the specified list.

https://bugzilla.redhat.com/show_bug.cgi?id=1356254