Fix missing languages in login screen language switcher
The list of available languages is pulled asynchronously, and when it's downloaded only after the language switcher select has been populated, angular updates the <select> element, but angular-patternfly 2.* won't update the visible selectpicker after.

Thus, sometimes, there are no languages in the selector.

This manually triggers a refresh of that selectpicker after the list is downloaded and the <select> is populated.

(This is not needed in master because angular-patternfly 3.* already does the right thing.)

https://bugzilla.redhat.com/show_bug.cgi?id=1363753