Allow approver to approve/deny service requests
This change adds a couple of buttons to the request details state to
allow for approving or denying the selected request. The buttons should
only be visible if the current user has the correct permissions and when
the `approval_state` for the request is `pending`.

The yarn.lock file is also updated here because for some reason the
dependencies are requiring two versions of node-sass.

Also cleans up a few linting issues.