Prevent selecting invalid warn options
When scheduling a retirement date in the future, a group of options are
available for selecting when a warning event should trigger (1, 2, 3, or
4) weeks prior the scheduled retirement date.

This change limits the selectable options to only values that make
sense. For example, you should not be able to select a warning 4 weeks
prior to retirement if you set retirement to tomorrow.