Add reset button to edit service modal
Include a reset button on the edit service modal to match the OpsUI. The
save and reset buttons are disabled unless the original values have been
changed.

Because just about every modal has the same actions at the footer of the
modal (cancel, reset, save), these buttons have been extracted to a
`modal-actions` component.

Also, since much of the functionality of each modal is very similar,
this change introduces a `BaseModalController` to pull up common
functionality.