Upgrade to latest angular-ui-bootstrap version
Upgrade to the most recent version of `angular-ui-bootstrap`. This
includes a significant number of bug fixes and improvements for a
library that we heavily use.

The biggest breaking change for the upgrade is the addition of a `uib`
prefix to all of `angular-ui-bootstrap` directives. This PR updates all
deprecated usage to the new version.

Because another dependency `angular-patternfly` is still using an older
version of `angular-ui-bootstrap`, a few minor things will not work
until the upgrade is complete there (about modal, toolbar, and tooltips).

A PR is open to make the changes there so hopefully it will be updated
in the next release https://github.com/patternfly/angular-patternfly/pull/353

https://github.com/angular-ui/bootstrap/wiki/Migration-guide-for-prefixes