Add .nvmrc and karma clear screen reporter
Add a `.nvmrc` file which is useful to keep developers in sync with
which node version is being used in production.

Add the `karma-clear-screen-reporter` which significantly cuts down on
the noise generated when running `gulp autotest`. Essentially on every
test run the screen is cleared so that you only see the report relevant
to your most recent changes.

Also update the documentation to mention these additions.