Various tweaks to allow js minification for prod build
Enable the `ng-strict-di` directive on `html` so that exceptions are
thrown when dependency injection is not handled properly.

Swap the babel plugin ng annotation tool with a loader variant. The
babel-focused version was missing several annotations which prevented
the build bundle from working properly.

Lock ngProgress to the latest version (released on GH but not npm). The
old version we are currently does not handle dependency injection
correctly and therefore cannot be minified. Also bump the explicit yarn
version to latest stable.

https://github.com/VictorBjelkholm/ngProgress/issues/116