Remove toastr as a global and import missing toastr css
Toastr messages do not display properly because the css associated with
the package is not being imported. This change removes Toastr as a
global (it is only used in the logger module) and imports the missing
css.

Also modify the eslintrc.json file to support ES6 module import/export
syntax.

Finally remove the Toasts factory as it is not used anywhere.