Add clear button to retire service modals
Adds a clear button to the bootstrap datepicker for the retire service
modals. This allows a user to remove a previously set retirement date.

Note that this will only work if you actually click the clear button, if
you simply delete the previous input and then click save, the previous
selected value from the datepicker will be submitted.

https://bugzilla.redhat.com/show_bug.cgi?id=1406944

A separate PR will be made for Euwe since there will likely be
conflicts.