Fix all broken pending specs
This change fixes all of our pending specs that were all failing from
the same issue. Previously the tests passed with the use of
`$rootScope.apply()` but stopped working dependencies were updated.

Using the `done()` callback to signal waiting for promises to resolve in
the tests causes all the pending specs to pass again.