Move all unit tests next to the units under test
Lots of test file movement to move away from a parallel hierarchy to the
co-located pattern that is recommend in future angular versions.

A few dead files were removed and a few other files were renamed to keep
naming conventions consistent.

Going forward, unit tests should be placed at the same level as the
tested file with an identical filename except with a `.spec.js`
extension.

https://www.pivotaltracker.com/story/show/137095739