Configure eslint to detect unused vars
Modify eslint configuration to error when a variable is unused. In some
situations it actually makes sense not to use select positional variable
arguments. In those scenarios you can prefix the variable with `_` to
hint eslint (and human readers) that the variable is intentionally not
being used.

The `angular/on-watch` rule was turned off because it conflicts with the
unused var rule by encouraging the use of variables in situations that
don't make sense.

Surprisingly, enabling this rule revealed several minor bugs related to
the use the wrong variables.