Add 'Last Message' to the requests explorer
Replace the 'ID' field with the 'Last Message' field in the requests
explorer. The id field is not used in any other collection view while
the message field generally provides more 'at a glance' information.

Because the message field is generally larger than the id column was,
shorten the approval status field to show only the icon and not the
text.

This change also modifies the other fields to have the stacked column
look, for consistency.

https://www.pivotaltracker.com/story/show/138616171
https://bugzilla.redhat.com/show_bug.cgi?id=1365253