Create new simpler TagList component for read-only tags
This change refactors the existing taggingWidget component into two
primary components: a smart TagEditor component that manages the state
of the tags and a reusable, presentational TagList component responsible
for rendering the tags for a resource in multiple contexts.

The new TagEditor will not need the extra 'Add' button and will instead
add tags to the TagList once a value is selected from a dropdown menu.

https://bugzilla.redhat.com/show_bug.cgi?id=1439634
https://www.pivotaltracker.com/story/show/143255995