Explicitly set window size for protractor before maximize
The pagination component is sensitive to screen size and at certain
dimensions will not render. For at least the MacOS/Chrome selenium
set-up the `window().maximize()` command does not work. The explicit
dimensions added should ensure that the automated test browser runs in
a window size close to how it is actually used.