Remove all traces of catalog list view and editing
Removes the Catalog list view along with associated component logic.
Also removes the CatalogEditorComponent, the dependent DualPaneSelector
and the associated state.

The catalogs mocks directory was also removed as the related specs are
no longer needed.

Currently, the catalog pagination and filter controls are a bit
unintuitive because they deal with catalogs and not catalog items while
each service catalog card is a representation of a catalog item.

https://www.pivotaltracker.com/story/show/143875995
https://www.pivotaltracker.com/story/show/143544747