pf-login - override patternfly login screen background image
With the upgrade from patternfly 3.45 to 3.46, the login screen background changed.
That's because patternfly made using a background image the default, but we don't have one so the default one gets used.

Overriding background-image for the login screen to achieve the original result.

Details in https://github.com/patternfly/patternfly/issues/1057