Timeline - remove onResize handler when destroying timeline
Fixes #1337

The timeline component listens for the resize event on window, and triggers a repaint of the timeline

But, it never removes that listener, even after moving to a screen without timeline,
causing a console error when opening the notification area (which generates a resize event).

    Error: <rect> attribute width: A negative value is not valid. ("-250")
    Error: <pattern> attribute width: A negative value is not valid. ("-250")
    Error: <rect> attribute width: A negative value is not valid. ("-250")
    Error: <rect> attribute width: A negative value is not valid. ("-250")

This fixes that, by removing the listener when the component is destroyed.