Session - use localStorage instead of sessionStorage
sessionStorage is not shared between tabs (unless one gets opened from the other),
and disappears when closing the tab

the same change was needed in ui-classic: https://github.com/ManageIQ/manageiq-ui-classic/pull/3065

The one place that's still using sessionStorage after this is the ApplianceInfo service, for caching. That looks more like ephemeral cache which doesn't have to be shared.