settings.help_menu.documentation - Fix missing href "Cannot read property 'match' of undefined"
The code assumes settings.help_menu.documentation contains a href, and fails without it.

But the href is there only if we want to override the link, fixing.

Also, the original code expects the href to always be http://api_location/api/http://the_href.com ,
now it will parse that if present, but fall back to just using the href if not.