Fix webmks console js/css paths
With `base: '/'`, the generated file contains...

    <base href="/">
    <link href="../styles/console_webmks-HASH.css" rel="stylesheet">
    ...
    <script src="../js/console_webmks-HASH.js">

that breaks on appliances where Service UI always runs as /ui/service.

The base causes all the relative links to be interporeted as /styles/... and /js/...,
instead of the proper relative paths (relative to `/ui/service/console/`), leading to 404s.

(The relative path in development is /console, so the base isn't needed for development either.)